Simplify ruff usage and tighten version constraint (#26)
* Simplify ruff CLI

Let ruff choose which files to look at. This can also be configured with
pyproject.toml entries following the docs
https://docs.astral.sh/ruff/configuration/

* Allow ruff to only update patch version number

Currently the versioning scheme for ruff includes breaking changes on
minor version updates and compatible changes with patch version updates.

Linting tools are often used in CI as a first pass, and pulling in a new
version with changes that find new issues can cause headaches. Pulling
in breaking changes and updates should be done carefully with manual
version bump.